@@ -2,19 +2,20 @@
   "name": "react-homework-template",
   "version": "0.1.0",
   "private": true,
-  "homepage": "https://ludmila121.github.io/goit-react-hw-04-feedback",
+  "homepage": "https://ludmila121.github.io/goit-react-hw-04-feedback/",
   "dependencies": {
+    "@emotion/react": "^11.11.0",
+    "@emotion/styled": "^11.11.0",
     "@testing-library/jest-dom": "^5.16.3",
     "@testing-library/react": "^12.1.4",
     "@testing-library/user-event": "^13.5.0",
     "node-sass": "^7.0.3",
     "normalize.css": "^8.0.1",
     "notiflix": "^3.2.6",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
-    "react-is": "^16.13.1",
+    "prop-types": "^15.8.1",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-scripts": "5.0.1",
-    "styled-components": "^5.3.10",
     "web-vitals": "^2.1.3"
   },
   "scripts": {
