@@ -2,10 +2,11 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import  App  from 'components/App';
 import './index.css';
-
+import reportWebVitals from './reportWebVitals'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
     <App />
   </React.StrictMode>
 );
+reportWebVitals();
